🎨 fix like button

+++ src/components/LikeButton.tsx
@@ -1,26 +1,33 @@
-import type { VFC } from "react";
-import { useState } from "react";
+import type { Dispatch, SetStateAction, VFC } from "react";
+import { useMemo } from "react";
 import { IoHeartOutline, IoHeartSharp } from "react-icons/io5";
 import { supabase } from "src/libs/supabase";
 
 type Props = {
   postId: string;
+  likes: string[];
+  setLikes: Dispatch<SetStateAction<string[]>>;
 };
 
-export const Like: VFC<Props> = (props) => {
-  const [isLike, setIsLike] = useState(false);
+export const LikeButton: VFC<Props> = (props) => {
+  const user = supabase.auth.user();
 
+  //いいねを取り消す
   const updateDisLike = async () => {
     try {
-      await supabase.from("likes").delete().eq("post_id", props.postId);
+      await supabase.from("likes").delete().eq("post_id", props.postId).eq("user_id", user?.id);
+      props.setLikes(
+        props.likes.filter((like) => {
+          return props.postId !== like;
+        })
+      );
     } catch (error) {
       console.error("error", error);
     }
   };
-
+  //いいねをつける
   const updateLike = async () => {
     try {
-      const user = supabase.auth.user();
       const updates = {
         user_id: user?.id,
         post_id: props.postId,
@@ -32,6 +39,7 @@ export const Like: VFC<Props> = (props) => {
       if (error) {
         throw error;
       }
+      props.setLikes([...props.likes, props.postId]);
     } catch (error) {
       console.error("error", error);
     }
@@ -39,19 +47,18 @@ export const Like: VFC<Props> = (props) => {
 
   const hadleClick = () => {
     isLike ? updateDisLike() : updateLike();
-    setIsLike((isLike) => {
-      return !isLike;
-    });
   };
+
+  const isLike = useMemo(() => {
+    return props.likes.some((like) => {
+      return props.postId === like;
+    });
+  }, [props.likes, props.postId]);
 
   return (
     <div className="flex justify-center">
       <button onClick={hadleClick}>
-        {isLike ? (
-          <IoHeartSharp size={22} color={"#5A5A5A"} className="" />
-        ) : (
-          <IoHeartOutline size={22} color={"#5A5A5A"} className="" />
-        )}
+        {isLike ? <IoHeartSharp size={22} color={"#5A5A5A"} /> : <IoHeartOutline size={22} color={"#5A5A5A"} />}
       </button>
     </div>
   );
